# openapi.json
{"openapi":"3.1.0","info":{"title":"RAG Service","version":"1.0.0"},"paths":{"/api/v1/auth/register":{"post":{"tags":["Authentication"],"summary":"Register","description":"Register a new user with email/username and password","operationId":"register_api_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","description":"Login with email/username and password","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/refresh":{"post":{"tags":["Authentication"],"summary":"Refresh Token","description":"Refresh access token using refresh token","operationId":"refresh_token_api_v1_auth_refresh_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["Authentication"],"summary":"Get Current User Info","description":"Get current user information","operationId":"get_current_user_info_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/auth/logout":{"post":{"tags":["Authentication"],"summary":"Logout","description":"Logout (invalidate token) - in production, add token to blacklist","operationId":"logout_api_v1_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/auth/google/login":{"get":{"tags":["Authentication"],"summary":"Google Login","description":"Initiate Google OAuth flow","operationId":"google_login_api_v1_auth_google_login_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/auth/google/callback":{"get":{"tags":["Authentication"],"summary":"Google Callback","description":"Handle Google OAuth callback","operationId":"google_callback_api_v1_auth_google_callback_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OAuthCallbackResponse"}}}}}}},"/api/v1/auth/github/login":{"get":{"tags":["Authentication"],"summary":"Github Login","description":"Initiate GitHub OAuth flow","operationId":"github_login_api_v1_auth_github_login_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/auth/github/callback":{"get":{"tags":["Authentication"],"summary":"Github Callback","description":"Handle GitHub OAuth callback","operationId":"github_callback_api_v1_auth_github_callback_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OAuthCallbackResponse"}}}}}}},"/api/v1/chats":{"post":{"tags":["Chats"],"summary":"Create Chat","description":"Create a new chat session","operationId":"create_chat_api_v1_chats_post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Chats"],"summary":"List Chats","description":"List user's chat sessions","operationId":"list_chats_api_v1_chats_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}},{"name":"include_deleted","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Include Deleted"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/chats/{chat_id}":{"get":{"tags":["Chats"],"summary":"Get Chat","description":"Get a specific chat with all messages","operationId":"get_chat_api_v1_chats__chat_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"chat_id","in":"path","required":true,"schema":{"type":"string","title":"Chat Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatWithMessages"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Chats"],"summary":"Update Chat","description":"Update chat title","operationId":"update_chat_api_v1_chats__chat_id__patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"chat_id","in":"path","required":true,"schema":{"type":"string","title":"Chat Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Chats"],"summary":"Delete Chat","description":"Soft delete a chat (can be restored)","operationId":"delete_chat_api_v1_chats__chat_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"chat_id","in":"path","required":true,"schema":{"type":"string","title":"Chat Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/chats/{chat_id}/restore":{"post":{"tags":["Chats"],"summary":"Restore Chat","description":"Restore a soft-deleted chat","operationId":"restore_chat_api_v1_chats__chat_id__restore_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"chat_id","in":"path","required":true,"schema":{"type":"string","title":"Chat Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/chats/{chat_id}/messages":{"post":{"tags":["Chats"],"summary":"Send Message","description":"Send a message and get RAG response","operationId":"send_message_api_v1_chats__chat_id__messages_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"chat_id","in":"path","required":true,"schema":{"type":"string","title":"Chat Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RAGQueryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admin/stats/system":{"get":{"tags":["Admin"],"summary":"Get System Stats","description":"Get overall system statistics","operationId":"get_system_stats_api_v1_admin_stats_system_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SystemStatsResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/admin/stats/rag":{"get":{"tags":["Admin"],"summary":"Get Rag Stats","description":"Get RAG engine statistics","operationId":"get_rag_stats_api_v1_admin_stats_rag_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RAGStatsResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/admin/users":{"get":{"tags":["Admin"],"summary":"List All Users","description":"List all users with statistics","operationId":"list_all_users_api_v1_admin_users_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":200,"minimum":1,"default":50,"title":"Limit"}},{"name":"search","in":"query","required":false,"schema":{"type":"string","title":"Search"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserStatsResponse"},"title":"Response List All Users Api V1 Admin Users Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admin/users/{user_id}":{"get":{"tags":["Admin"],"summary":"Get User Details","description":"Get detailed user information","operationId":"get_user_details_api_v1_admin_users__user_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserStatsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Admin"],"summary":"Update User Settings","description":"Update user settings (admin only)","operationId":"update_user_settings_api_v1_admin_users__user_id__patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminUserUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserStatsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admin/users/{user_id}/conversations":{"get":{"tags":["Admin"],"summary":"Export User Conversations","description":"Export all conversations for a user (including deleted)","operationId":"export_user_conversations_api_v1_admin_users__user_id__conversations_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ConversationExport"},"title":"Response Export User Conversations Api V1 Admin Users  User Id  Conversations Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admin/conversations/{chat_id}":{"get":{"tags":["Admin"],"summary":"Export Conversation","description":"Export a specific conversation","operationId":"export_conversation_api_v1_admin_conversations__chat_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"chat_id","in":"path","required":true,"schema":{"type":"string","title":"Chat Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConversationExport"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Root","description":"Root endpoint","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AdminUserUpdate":{"properties":{"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active"},"is_admin":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Admin"},"max_messages_per_day":{"anyOf":[{"type":"integer","maximum":10000.0,"minimum":0.0},{"type":"null"}],"title":"Max Messages Per Day"},"rate_limit_per_minute":{"anyOf":[{"type":"integer","maximum":1000.0,"minimum":1.0},{"type":"null"}],"title":"Rate Limit Per Minute"}},"type":"object","title":"AdminUserUpdate","description":"Admin can update user settings"},"AuthProvider":{"type":"string","enum":["local","google","github"],"title":"AuthProvider","description":"Authentication provider types"},"ChatCreate":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title","default":"New Chat"}},"type":"object","title":"ChatCreate"},"ChatListResponse":{"properties":{"total":{"type":"integer","title":"Total"},"chats":{"items":{"$ref":"#/components/schemas/ChatResponse"},"type":"array","title":"Chats"},"skip":{"type":"integer","title":"Skip"},"limit":{"type":"integer","title":"Limit"}},"type":"object","required":["total","chats","skip","limit"],"title":"ChatListResponse"},"ChatResponse":{"properties":{"id":{"type":"string","title":"Id"},"user_id":{"type":"string","title":"User Id"},"title":{"type":"string","title":"Title"},"is_deleted":{"type":"boolean","title":"Is Deleted"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"message_count":{"type":"integer","title":"Message Count","default":0},"last_message_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Message At"}},"type":"object","required":["id","user_id","title","is_deleted","created_at","updated_at"],"title":"ChatResponse"},"ChatUpdate":{"properties":{"title":{"type":"string","maxLength":255,"minLength":1,"title":"Title"}},"type":"object","required":["title"],"title":"ChatUpdate"},"ChatWithMessages":{"properties":{"id":{"type":"string","title":"Id"},"user_id":{"type":"string","title":"User Id"},"title":{"type":"string","title":"Title"},"is_deleted":{"type":"boolean","title":"Is Deleted"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"message_count":{"type":"integer","title":"Message Count","default":0},"last_message_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Message At"},"messages":{"items":{"$ref":"#/components/schemas/MessageResponse"},"type":"array","title":"Messages","default":[]}},"type":"object","required":["id","user_id","title","is_deleted","created_at","updated_at"],"title":"ChatWithMessages"},"ConversationExport":{"properties":{"user_id":{"type":"string","title":"User Id"},"chat_id":{"type":"string","title":"Chat Id"},"title":{"type":"string","title":"Title"},"messages":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Messages"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"message_count":{"type":"integer","title":"Message Count"}},"type":"object","required":["user_id","chat_id","title","messages","created_at","message_count"],"title":"ConversationExport","description":"Export user conversations"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MessageCreate":{"properties":{"content":{"type":"string","maxLength":10000,"minLength":1,"title":"Content"}},"type":"object","required":["content"],"title":"MessageCreate"},"MessageResponse":{"properties":{"id":{"type":"string","title":"Id"},"chat_session_id":{"type":"string","title":"Chat Session Id"},"role":{"type":"string","title":"Role"},"content":{"type":"string","title":"Content"},"order_index":{"type":"integer","title":"Order Index"},"metadata":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Metadata"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","chat_session_id","role","content","order_index","created_at"],"title":"MessageResponse"},"OAuthCallbackResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"type":"string","title":"Refresh Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"user":{"additionalProperties":true,"type":"object","title":"User"}},"type":"object","required":["access_token","refresh_token","user"],"title":"OAuthCallbackResponse"},"RAGQueryResponse":{"properties":{"message_id":{"type":"string","title":"Message Id"},"chat_id":{"type":"string","title":"Chat Id"},"user_message":{"$ref":"#/components/schemas/MessageResponse"},"assistant_message":{"$ref":"#/components/schemas/MessageResponse"},"processing_time_ms":{"type":"number","title":"Processing Time Ms"}},"type":"object","required":["message_id","chat_id","user_message","assistant_message","processing_time_ms"],"title":"RAGQueryResponse","description":"Response from RAG query"},"RAGStatsResponse":{"properties":{"model":{"type":"string","title":"Model"},"index":{"type":"string","title":"Index"},"documents_count":{"type":"integer","title":"Documents Count"},"device":{"type":"string","title":"Device"},"status":{"type":"string","title":"Status"}},"type":"object","required":["model","index","documents_count","device","status"],"title":"RAGStatsResponse","description":"RAG engine statistics"},"RefreshTokenRequest":{"properties":{"refresh_token":{"type":"string","title":"Refresh Token"}},"type":"object","required":["refresh_token"],"title":"RefreshTokenRequest"},"SystemStatsResponse":{"properties":{"total_users":{"type":"integer","title":"Total Users"},"active_users":{"type":"integer","title":"Active Users"},"total_chats":{"type":"integer","title":"Total Chats"},"total_messages":{"type":"integer","title":"Total Messages"},"messages_today":{"type":"integer","title":"Messages Today"},"rag_engine_status":{"type":"string","title":"Rag Engine Status"},"documents_indexed":{"type":"integer","title":"Documents Indexed"},"average_response_time_ms":{"type":"number","title":"Average Response Time Ms"}},"type":"object","required":["total_users","active_users","total_chats","total_messages","messages_today","rag_engine_status","documents_indexed","average_response_time_ms"],"title":"SystemStatsResponse","description":"Overall system statistics"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"type":"string","title":"Refresh Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token","refresh_token"],"title":"Token"},"UserCreate":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username"},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"password":{"type":"string","minLength":8,"title":"Password"}},"type":"object","required":["email","password"],"title":"UserCreate"},"UserLogin":{"properties":{"login":{"type":"string","title":"Login"},"password":{"type":"string","title":"Password"}},"type":"object","required":["login","password"],"title":"UserLogin"},"UserResponse":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username"},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"id":{"type":"string","title":"Id"},"auth_provider":{"$ref":"#/components/schemas/AuthProvider"},"is_active":{"type":"boolean","title":"Is Active"},"is_admin":{"type":"boolean","title":"Is Admin"},"is_verified":{"type":"boolean","title":"Is Verified"},"avatar_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Avatar Url"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"last_login_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Login At"},"max_messages_per_day":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Messages Per Day"},"rate_limit_per_minute":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Rate Limit Per Minute"}},"type":"object","required":["email","id","auth_provider","is_active","is_admin","is_verified","created_at"],"title":"UserResponse"},"UserStatsResponse":{"properties":{"id":{"type":"string","title":"Id"},"email":{"type":"string","title":"Email"},"username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username"},"auth_provider":{"type":"string","title":"Auth Provider"},"is_active":{"type":"boolean","title":"Is Active"},"is_admin":{"type":"boolean","title":"Is Admin"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"last_login_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Login At"},"total_chats":{"type":"integer","title":"Total Chats"},"total_messages":{"type":"integer","title":"Total Messages"},"messages_today":{"type":"integer","title":"Messages Today"},"max_messages_per_day":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Messages Per Day"},"rate_limit_per_minute":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Rate Limit Per Minute"}},"type":"object","required":["id","email","username","auth_provider","is_active","is_admin","created_at","last_login_at","total_chats","total_messages","messages_today","max_messages_per_day","rate_limit_per_minute"],"title":"UserStatsResponse","description":"Detailed user statistics for admin"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}